// web/src/api/generated/models/reading-quiz.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChapterId } from './chapter-id';
// May contain unused imports in some cases
// @ts-ignore
import type { QuizQuestion } from './quiz-question';
// May contain unused imports in some cases
// @ts-ignore
import type { ReadingQuizId } from './reading-quiz-id';

export interface ReadingQuiz {
    'id'?: ReadingQuizId;
    'chapterId'?: ChapterId;
    'questions'?: Array<QuizQuestion>;
    'aiModel'?: string;
    'createTime'?: string;
    'questionCount'?: number;
}