// web/src/api/generated/models/quiz-question.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface QuizQuestion {
    'type'?: QuizQuestionTypeEnum;
    'difficulty'?: QuizQuestionDifficultyEnum;
    'question'?: string;
    'options'?: Array<string>;
    'correctAnswer'?: string;
    'explanation'?: string;
}

export const QuizQuestionTypeEnum = {
    CHOICE: 'CHOICE',
    FILL: 'FILL',
    TRUE_FALSE: 'TRUE_FALSE',
    SHORT_ANSWER: 'SHORT_ANSWER'
} as const;

export type QuizQuestionTypeEnum = typeof QuizQuestionTypeEnum[keyof typeof QuizQuestionTypeEnum];
export const QuizQuestionDifficultyEnum = {
    EASY: 'EASY',
    MEDIUM: 'MEDIUM',
    HARD: 'HARD'
} as const;

export type QuizQuestionDifficultyEnum = typeof QuizQuestionDifficultyEnum[keyof typeof QuizQuestionDifficultyEnum];